nit: rename py_ecc.bls.g2_primatives and add an alias
